grok_requirements.txt

requests>=2.25.1
openai>=1.0.0
python-dotenv>=0.19.0
